Rebasing around rwill's modifications

+++ source/r_edge.c
@@ -462,7 +462,7 @@ void R_LeadingEdge (edge_t *edge)
 	espan_t			*span;
 	surf_t			*surf, *surf2;
 	int				iu;
-	double			fu, newzi, testzi, newzitop, newzibottom;
+	float			fu, newzi, testzi, newzitop, newzibottom;
 
 	if (edge->surfs[1])
 	{
